+++ SandboxiePlus/SandMan/OnlineUpdater.cpp
@@ -446,6 +446,7 @@ bool COnlineUpdater::HandleUpdate()
 {
 	QString PendingUpdate;
 
+	QString OnNewRelease = GetOnNewReleaseOption();
 	bool bNewRelease = false;
 	QVariantMap Release = m_UpdateData["release"].toMap();
 	QString ReleaseStr = Release["version"].toString();
@@ -457,7 +458,6 @@ bool COnlineUpdater::HandleUpdate()
 	}
 
 	QString OnNewUpdate = GetOnNewUpdateOption();
-
 	bool bNewUpdate = false;
 	QVariantMap Update = m_UpdateData["update"].toMap();
 	QString UpdateStr = Update["version"].toString();
@@ -486,7 +486,16 @@ bool COnlineUpdater::HandleUpdate()
 	// solution: apply updates silently, then prompt to install new release, else prioritize installing new releases over updating the existing one
 	//
 
-	QString OnNewRelease = GetOnNewReleaseOption();
+	bool bAllowAuto = g_CertInfo.active && !g_CertInfo.expired; // To use automatic updates a valid certificate is required
+
+	//
+	// if we allow for version updates but not for automatic instalation/download of new release 
+	// ignore the release and install it using the version updater
+	//
+
+	if (bNewUpdate && bNewRelease && !bAllowAuto)
+		bNewRelease = false;
+
 	bool bCanRunInstaller = (m_CheckMode == eAuto && OnNewRelease == "install");
 	bool bIsInstallerReady = false;
 	if (bNewRelease) 
@@ -502,7 +511,7 @@ bool COnlineUpdater::HandleUpdate()
 			// clear when not up to date
 			theConf->DelValue("Updater/InstallerVersion");
 
-			if ((bCanRunInstaller || (m_CheckMode == eAuto && OnNewRelease == "download")) || AskDownload(Release))
+			if ((bAllowAuto && (bCanRunInstaller || (m_CheckMode == eAuto && OnNewRelease == "download"))) || AskDownload(Release, bAllowAuto))
 			{
 				if (DownloadInstaller(Release, m_CheckMode == eManual))
 					return true;
@@ -524,7 +533,7 @@ bool COnlineUpdater::HandleUpdate()
 				// clear when not up to date
 				theConf->DelValue("Updater/UpdateVersion");
 
-				if ((bCanApplyUpdate || (m_CheckMode == eAuto && OnNewUpdate == "download")) || AskDownload(Update))
+				if ((bCanApplyUpdate || (m_CheckMode == eAuto && OnNewUpdate == "download")) || AskDownload(Update, true))
 				{
 					if (DownloadUpdate(Update, m_CheckMode == eManual))
 						return true;
@@ -555,7 +564,7 @@ bool COnlineUpdater::HandleUpdate()
 	return bNewRelease || bNewUpdate;
 }
 
-bool COnlineUpdater::AskDownload(const QVariantMap& Data)
+bool COnlineUpdater::AskDownload(const QVariantMap& Data, bool bAuto)
 {
 	QString VersionStr = MakeVersionStr(Data);
 
@@ -568,12 +577,25 @@ bool COnlineUpdater::AskDownload(const QVariantMap& Data)
 	QVariantMap Installer = Data["installer"].toMap();
 	QString DownloadUrl = Installer["downloadUrl"].toString();
 
-	if (!DownloadUrl.isEmpty())
+	enum EAction
+	{
+		eNone = 0,
+		eDownload,
+		eNotify,
+	} Action = eNone;
+
+	if (bAuto && !DownloadUrl.isEmpty()) {
+		Action = eDownload;
 		FullMessage += tr("<p>Do you want to download the installer?</p>");
-	else if(Data.contains("files"))
+	}
+	else if (bAuto && Data.contains("files")) {
+		Action = eDownload;
 		FullMessage += tr("<p>Do you want to download the updates?</p>");
-	else if (!UpdateUrl.isEmpty())
-		FullMessage += tr("<p>Do you want to go to the <a href=\"%1\">update page</a>?</p>").arg(UpdateUrl);
+	}
+	else if (!UpdateUrl.isEmpty()) {
+		Action = eNotify;
+		FullMessage += tr("<p>Do you want to go to the <a href=\"%1\">download page</a>?</p>").arg(UpdateUrl);
+	}
 
 	CCheckableMessageBox mb(theGUI);
 	mb.setWindowTitle("Sandboxie-Plus");
@@ -584,18 +606,18 @@ bool COnlineUpdater::AskDownload(const QVariantMap& Data)
 	mb.setCheckBoxText(tr("Don't show this update anymore."));
 	mb.setCheckBoxVisible(m_CheckMode != eManual);
 
-	if (!UpdateUrl.isEmpty() || !DownloadUrl.isEmpty() || Data.contains("files")) {
+	if (Action != eNone) {
 		mb.setStandardButtons(QDialogButtonBox::Yes | QDialogButtonBox::No | QDialogButtonBox::Cancel);
 		mb.setDefaultButton(QDialogButtonBox::Yes);
-	}
-	else
+	} else
 		mb.setStandardButtons(QDialogButtonBox::Ok);
 
 	mb.exec();
 
 	if (mb.clickedStandardButton() == QDialogButtonBox::Yes)
 	{
-		if (!DownloadUrl.isEmpty() || Data.contains("files")) {
+		if (Action == eDownload) 
+		{
 			m_CheckMode = eManual;
 			return true;
 		}
@@ -604,7 +626,8 @@ bool COnlineUpdater::AskDownload(const QVariantMap& Data)
 	}
 	else 
 	{
-		if (mb.clickedStandardButton() == QDialogButtonBox::Cancel) {
+		if (mb.clickedStandardButton() == QDialogButtonBox::Cancel) 
+		{
 			theConf->SetValue("Updater/PendingUpdate", ""); 
 			theGUI->UpdateLabel();
 		}