+++ CHANGELOG.md
@@ -10,6 +10,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - fixed running sandboxed processes located in a imdisk volume [#3472](https://github.com/sandboxie-plus/Sandboxie/discussions/3472)
 
+### Changed
+- without an active, non expired, supporter certificate, automatic updates/downloads are not longer available for the stable channel
+ - the autoamtic updater will still work and notify about new stable releases, the user will be guided to visit the download page and download the latest installer manually
+
 
 
 ## [1.12.3 / 5.67.3] - 2023-12-02